LPS-130793 account-rest-impl: fixes compile errors
